backend/requirements.txt
celery==5.5.3
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pillow==11.3.0